Move all banners to `/elements/banners/`
