+++ src/app/components/elements/banners/CoronavirusWarningBanner.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {Alert, Container} from "reactstrap";
-import {SITE_TITLE} from "../../services";
+import {SITE_TITLE} from "../../../services";
 
 // Before deleting, it might be worth altering this to be more generic and controlled by content on ElasticSearch
 export const CoronavirusWarningBanner = () => (
