+++ src/app/components/elements/banners/IntendedAudienceWarningBanner.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import {ContentBaseDTO} from "../../../IsaacApiTypes";
-import {examBoardLabelMap, isAda, isIntendedAudience, isLoggedIn, stageLabelMap, useUserViewingContext} from "../../services";
-import {selectors, useAppSelector} from "../../state";
-import {RenderNothing} from "../elements/RenderNothing";
+import {ContentBaseDTO} from "../../../../IsaacApiTypes";
+import {examBoardLabelMap, isAda, isIntendedAudience, isLoggedIn, stageLabelMap, useUserViewingContext} from "../../../services";
+import {selectors, useAppSelector} from "../../../state";
+import {RenderNothing} from "../RenderNothing";
 import { Link } from "react-router-dom";
 import { Alert } from "reactstrap";
