+++ src/app/components/elements/banners/RevisionWarningBanner.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import {AppState, selectors, useAppSelector} from "../../state";
-import {RenderNothing} from "../elements/RenderNothing";
+import {AppState, selectors, useAppSelector} from "../../../state";
+import {RenderNothing} from "../RenderNothing";
 import { Alert } from "reactstrap";
 
 export function RevisionWarningBanner() {
