+++ src/app/components/navigation/IsaacApp.tsx
@@ -45,8 +45,8 @@ import {AuthError} from "../pages/AuthError";
 import {SessionExpired} from "../pages/SessionExpired";
 import {ConsistencyError} from "../pages/ConsistencyError";
 import {Search} from "../pages/Search";
-import {ResearchNotificationBanner} from "./ResearchNotificationBanner";
-import {EmailVerificationBanner} from "./EmailVerificationBanner";
+import {ResearchNotificationBanner} from "../elements/banners/ResearchNotificationBanner";
+import {EmailVerificationBanner} from "../elements/banners/EmailVerificationBanner";
 import {Toasts} from "./Toasts";
 import {AdminUserManager} from "../pages/AdminUserManager";
 import {AdminStats} from "../pages/AdminStats";
@@ -62,7 +62,7 @@ import {FreeTextBuilder} from "../pages/FreeTextBuilder";
 import {MarkdownBuilder} from "../pages/MarkdownBuilder";
 import SiteSpecific from "../site/siteSpecificComponents";
 import {surveyNotificationModal} from "../elements/modals/SurveyNotificationModal";
-import {DowntimeWarningBanner} from "./DowntimeWarningBanner";
+import {DowntimeWarningBanner} from "../elements/banners/DowntimeWarningBanner";
 import {ErrorBoundary} from "react-error-boundary";
 import {ChunkOrClientError} from "../pages/ClientError";
 import {Loading} from "../handlers/IsaacSpinner";
@@ -74,7 +74,7 @@ import {QuestionFinder} from "../pages/QuestionFinder";
 import {SessionCookieExpired} from "../pages/SessionCookieExpired";
 import { AccountDeletion } from '../pages/AccountDeletion';
 import { AccountDeletionSuccess } from '../pages/AccountDeletionSuccess';
-import { IsaacScienceLaunchBanner } from './IsaacScienceLaunchBanner';
+import { IsaacScienceLaunchBanner } from '../elements/banners/IsaacScienceLaunchBanner';
 import { RequireAuth } from './UserAuthentication';
 import { FigureNumberingProvider } from '../elements/FigureNumberingProvider';
 import { QualtricsRedirect } from './external/QualtricsRedirect';
