+++ src/app/components/pages/Concept.tsx
@@ -11,7 +11,7 @@ import {WithFigureNumbering} from "../elements/WithFigureNumbering";
 import {TitleAndBreadcrumb} from "../elements/TitleAndBreadcrumb";
 import {NavigationLinks} from "../elements/NavigationLinks";
 import {Markup} from "../elements/markup";
-import {IntendedAudienceWarningBanner} from "../navigation/IntendedAudienceWarningBanner";
+import {IntendedAudienceWarningBanner} from "../elements/banners/IntendedAudienceWarningBanner";
 import {SupersededDeprecatedStandaloneContentWarning} from "../navigation/SupersededDeprecatedWarning";
 import {CanonicalHrefElement} from "../navigation/CanonicalHrefElement";
 import {MetaDescription} from "../elements/MetaDescription";
@@ -22,7 +22,7 @@ import { ShowLoadingQuery } from "../handlers/ShowLoadingQuery";
 import { NotFound } from "./NotFound";
 import { PageMetadata } from "../elements/PageMetadata";
 import { getAccessibilityTags, useAccessibilitySettings } from "../../services/accessibility";
-import { InaccessibleContentWarningBanner } from "../navigation/InaccessibleContentWarningBanner";
+import { InaccessibleContentWarningBanner } from "../elements/banners/InaccessibleContentWarningBanner";
 import { skipToken } from "@reduxjs/toolkit/query";
 import { GameboardContentSidebar } from "../elements/sidebar/GameboardContentSidebar";
 import { ConceptSidebar } from "../elements/sidebar/RelatedContentSidebar";
